Beginning of FsCheck.Xunit CoreClr support
This is an attempt to get FsCheck.Xunit working against the core clr. A
new portable profile for FsCheck.Xunit has been created and its output
added to the paket template for the FsCheck.Xunit package. The compiled
package has been tested with partial success - tests can be created and
run with the PropertyAttribute, but the test runner doesn't display
timing information for Properties.